+++ global.json
@@ -1,3 +1,8 @@
 {
-  "projects": [ "src" ]
+  "projects": [
+    "src"
+  ],
+  "sdk": {
+    "version": "1.0.0-preview2-1-003177"
+  }
 }
