+++ samples/LocalizationSample/Startup.cs
@@ -18,12 +18,10 @@ namespace LocalizationSample
     {
         public void ConfigureServices(IServiceCollection services)
         {
-            services.AddLocalization(options => options.ResourcesPath = "POFiles");
-            services.AddTransient<IStringLocalizerFactory, POStringLocalizerFactory>();
+            services.AddLocalization(options => options.ResourcesPath = "My/Resources");
         }
 
-        public void Configure(IApplicationBuilder app, IStringLocalizer<Startup> SR,
-            IStringLocalizerFactory factory)
+        public void Configure(IApplicationBuilder app, IStringLocalizer<Startup> SR)
         {
             var supportedCultures = new List<CultureInfo>
             {
@@ -56,10 +54,6 @@ namespace LocalizationSample
 
             app.Use(async (context, next) =>
             {
-                CultureInfo.CurrentUICulture = new CultureInfo("es");
-                var suse = factory.Create(typeof(Startup));
-                var result = suse["CLOSE"];
-
                 if (context.Request.Path.Value.EndsWith("favicon.ico"))
                 {
                     // Pesky browsers
