+++ src/Microsoft.Extensions.Localization/Internal/POLines/CommentBase.cs
@@ -1,0 +1,12 @@
+// Copyright (c) .NET Foundation. All rights reserved. 
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information. 
+
+using System;
+using System.Text;
+
+namespace Microsoft.Extensions.Localization.Internal.POLines
+{
+    public abstract class CommentBase : TokenLine
+    {
+    }
+}
